new network driver: `pasta` (with port driver `implicit`)
Pasta: https://passt.top/passt/
Usage: `rootlesskit --net=pasta --port-driver=implicit`

- No support for explicit port forwarding (`rootlessctl add-ports`),
  as pasta doesn't support it yet.
  Use `--port-driver=implicit` to let pasta forward TCP ports implicitly.
  The forwarded ports are not visible in `rootlessctl list-ports`.

- No support for forwarding UDP ports

- Tested with pasta 2023_06_25.32660ce on Ubuntu 23.04.
  Doesn't work with 2023_06_03.429e1a7: `Option --no-copy-routes needs --config-net`
  (This is printed despite that `--no-copy-routes` is not specified)

- Doesn't work with Ubuntu 23.04's dpkg (passt_0.0~git20230216.4663ccc-1_amd64.deb):
  `Couldn't open user namespace /proc/51813/ns/user: Permission denied`
  Likely to be related to AppArmor.

+++ hack/integration-docker.sh
@@ -20,10 +20,14 @@ $CURL "http://127.0.0.1:8080"
 $CURL "http://${nonloopback}:8080" && ( ERROR "should fail"; exit 1 )
 docker rm -f nginx
 
-docker run -d --name=nginx -p "${nonloopback}:8080:80" nginx:alpine
-sleep 2
-$CURL "http://127.0.0.1:8080" && ( ERROR "should fail"; exit 1 )
-$CURL "http://${nonloopback}:8080"
-docker rm -f nginx
+# port driver "implicit" does not support binding on a non-loopback address:
+# "Error starting userland proxy: listen tcp4 172.18.0.2:8080: bind: cannot assign requested address."
+if [ "$DOCKERD_ROOTLESS_ROOTLESSKIT_PORT_DRIVER" != "implicit" ]; then
+	docker run -d --name=nginx -p "${nonloopback}:8080:80" nginx:alpine
+	sleep 2
+	$CURL "http://127.0.0.1:8080" && ( ERROR "should fail"; exit 1 )
+	$CURL "http://${nonloopback}:8080"
+	docker rm -f nginx
+fi
 
 INFO "===== PASSING ====="
